Numeric.Algebra.Normed

Description

Provides the Normed typeclass.

Since: 0.1

Documentation

class Normed s where

Defines a "norm" on a given structure. We have the following laws:

Positive definiteness: norm x > 0 for x /= 0, norm 0 = 0.

Subadditivity: norm (x + y) <= norm x + norm y.

Inversion : norm x = norm (-x).

Sign : sgn x = x / norm x.

Notice these laws require various algebraic assumptions, from a semigroup up to a group structure. It is expected that types that implement the relevant typeclasses will follow the respective laws.

We do not require any of these typeclasses as constraints as it is possible that we may want to abstract over "positive" types with a type that does not have an additive semigroup instance (consider multiplicative groups).

Since: 0.1

Methods

norm :: s -> s

Returns the norm.

Since: 0.1

sgn :: s -> s

Sign function. Notice that when norm is abs, this is the familiar signum function i.e.

sgn x
  | x < 0 = -1
  | x == 0 = 0
  | x > 0 = 1

Since: 0.1