Numeric.Convert.Internal

Description

Since: 0.1

Integers

class FromInteger a where

Replaces base's fromInteger functionality for when we do not have a Num instance.

1_000 :: Num a => a

-- becomes

fromZ 1_000 :: FromInteger a => a

Note that fromInteger's deficiencies are inherited e.g. Natural is partial, bounded types have over/underflow issues.

Since: 0.1

Methods

fromZ :: Integer -> a

Since: 0.1

class ToRational a => ToInteger a where

Integer embedding.

Since: 0.1

Methods

toZ :: a -> Integer

Since: 0.1

Rationals

class FromInteger a => FromRational a where

Replaces base's fromRational functionality for when we do not have a Fractional instance.

5.5 :: Fractional a => a

-- becomes

fromQ 5.5 :: FromRational a => a

Note that fromRational's deficiencies are inherited e.g. Natural is partial, bounded types have over/underflow issues.

Since: 0.1

Methods

fromQ :: Rational -> a

Since: 0.1

class ToReal a => ToRational a where

Rational embedding.

Since: 0.1

Methods

toQ :: a -> Rational

Since: 0.1

Reals

class FromRational a => FromReal a where

Conversion from Double.

Since: 0.1

Methods

fromR :: Double -> a

Since: 0.1

class ToReal a where

Conversion to Double.

Since: 0.1

Methods

toR :: a -> Double

Since: 0.1